+++ components/indicator-visualizations/area/index.ts
