+++ components/indicator-visualizations/bar/component.tsx
@@ -42,21 +42,19 @@ const Chart: FC<ChartProps> = ({ widgetData, widgetConfig }: ChartProps) => {
     ...rest
   } = widgetConfig;
   return (
-    <div className="py-10">
-      <ResponsiveContainer width="100%" height={500} {...rest}>
-        <BarChart width={400} height={500} data={widgetData} {...rest}>
-          {cartesianGrid && (<CartesianGrid {...cartesianGrid} />)}
-          {cartesianAxis && (<CartesianAxis {...cartesianAxis} />)}
-          {xAxis && (<XAxis {...xAxis} />)}
-          {yAxis && (<YAxis {...yAxis} />)}
-          {bars && (
-            Object.keys(bars)
-              .map((bar, index) => (<Bar key={bar} {...bars[bar]} fill={colors[index]} />
-              )))}
-          {tooltip && (<Tooltip />)}
-        </BarChart>
-      </ResponsiveContainer>
-    </div>
+    <ResponsiveContainer width="100%" height={500} {...rest}>
+      <BarChart width={400} height={500} data={widgetData} {...rest}>
+        {cartesianGrid && (<CartesianGrid {...cartesianGrid} />)}
+        {cartesianAxis && (<CartesianAxis {...cartesianAxis} />)}
+        {xAxis && (<XAxis {...xAxis} />)}
+        {yAxis && (<YAxis {...yAxis} />)}
+        {bars && (
+          Object.keys(bars)
+            .map((bar, index) => (<Bar key={bar} {...bars[bar]} fill={colors[index]} />
+            )))}
+        {tooltip && (<Tooltip />)}
+      </BarChart>
+    </ResponsiveContainer>
   );
 };
