+++ components/indicator-visualizations/bar/index.ts
