+++ components/indicator-visualizations/line/component.tsx
@@ -47,20 +47,18 @@ const Chart: FC<ChartProps> = ({ widgetData, widgetConfig }: ChartProps) => {
   } = widgetConfig;
 
   return (
-    <div className="py-10">
-      <ResponsiveContainer width="100%" height={500}>
-        <LineChart width={400} height={200} data={widgetData}>
-          {cartesianGrid && (<CartesianGrid {...cartesianGrid} />)}
-          {cartesianAxis && (<CartesianAxis {...cartesianAxis} />)}
-          {xAxis && (<XAxis {...xAxis} />)}
-          {yAxis && (<YAxis {...yAxis} />)}
-          {lines && Object.keys(lines).map((line, index) => (
-            <Line key={line} {...lines[line]} stroke={colors[index]} />
-          ))}
-          {tooltip && (<Tooltip />)}
-        </LineChart>
-      </ResponsiveContainer>
-    </div>
+    <ResponsiveContainer width="100%" height={500}>
+      <LineChart width={400} height={200} data={widgetData}>
+        {cartesianGrid && (<CartesianGrid {...cartesianGrid} />)}
+        {cartesianAxis && (<CartesianAxis {...cartesianAxis} />)}
+        {xAxis && (<XAxis {...xAxis} />)}
+        {yAxis && (<YAxis {...yAxis} />)}
+        {lines && Object.keys(lines).map((line, index) => (
+          <Line key={line} {...lines[line]} stroke={colors[index]} />
+        ))}
+        {tooltip && (<Tooltip />)}
+      </LineChart>
+    </ResponsiveContainer>
   );
 };
