+++ components/indicator-visualizations/map/constants.js
