+++ components/indicator-visualizations/map/legend/index.ts
