+++ components/indicator-visualizations/map/zoom/component.tsx
