+++ components/indicator-visualizations/map/zoom/index.ts
