+++ components/indicator-visualizations/table/component.tsx
