+++ components/indicator-visualizations/table/index.ts
